server/requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
dnspython==1.16.0
ecdsa==0.15
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
gevent==1.4.0
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
monotonic==1.5
opentok==2.10.0
pkg-resources==0.0.0
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==19.1.0
python-dotenv==0.12.0
python-engineio==3.12.1
python-jose==3.1.0
python-socketio==4.5.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
